x64 warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@2716 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TrafficCounter/src/commonheaders.h
@@ -149,7 +149,7 @@ union
 
 PROTOLIST *ProtoList; // Данные обо всех аккаунтах.
 PROTOLIST OverallInfo; // Суммарные данные по видимым аккаунтам.
-BYTE NumberOfAccounts;
+int NumberOfAccounts;
 WORD Stat_SelAcc; // Выбранные аккаунты в окне статистики
 HWND TrafficHwnd;
 DWORD mirandaVer;
