+++ plugins/TrafficCounter/src/vars.h
@@ -2,7 +2,7 @@
 #define _vars_h
 
 //extern BYTE numProto;
-extern BYTE NumberOfAccounts;
+extern int NumberOfAccounts;
 //extern PROTOLIST *ProtoList;
 void RegisterVariablesTokens(void);
